Interim commit: working on documentation, and fixing directory and page structures
