+++ src/ToMA.py
@@ -31,7 +31,7 @@ import streamlit as st
 
 
 st.header("ToMA: Topological Material Analysis")
-st.markdown("ToMA is designed to help researchers analyse the structure of materials using tools from geometry and topology. Currently, it uses (Oineus)[https://github.com/anigmetov/oineus] and (Diode) under the hood to take care of some calculations and processing of the data. As such, it is currently licensed under a GNU GPL license, which you can view from the 'License' page in the side bar.")
+st.markdown("ToMA is designed to help researchers analyse the structure of materials using tools from geometry and topology. Currently, it uses [Oineus](https://github.com/anigmetov/oineus) and [Diode](https://github.com/mrzv/diode) under the hood to take care of some calculations and processing of the data. As such, it is currently licensed under a GNU GPL license, which you can view from the [License](/License) page in the side bar.")
 
 
 # welcome_page = st.Page("welcome.py", title="ToMA")
